refactor(SCALE): extract SCALE en/decoding exceptions as custom RuntimeExceptions
also, move the existing ScaleEncodingException into `utils`
