+++ src/main/java/com/limechain/network/protocol/transactions/TransactionsEngine.java
@@ -3,7 +3,7 @@ package com.limechain.network.protocol.transactions;
 import com.limechain.network.ConnectionManager;
 import com.limechain.network.protocol.transactions.scale.TransactionsReader;
 import com.limechain.network.protocol.transactions.scale.TransactionsWriter;
-import com.limechain.network.protocol.warp.exception.ScaleEncodingException;
+import com.limechain.utils.scale.exceptions.ScaleEncodingException;
 import com.limechain.sync.warpsync.SyncedState;
 import io.emeraldpay.polkaj.scale.ScaleCodecReader;
 import io.emeraldpay.polkaj.scale.ScaleCodecWriter;
